oracle-monitor/collectors/queue_collector/kafka_collector.py
from typing import Dict, Any, List
from kafka import KafkaAdminClient
from kafka.cluster import ClusterMetadata
from collectors.base.collector import Collector
from collectors.base.config import Config

class QueueCollector(Collector):
    def __init__(self):
        self.bootstrap_servers = Config.KAFKA_BOOTSTRAP_SERVERS
        
    def collect(self) -> List[Dict[str, Any]]:
        queues = []
        try:
            # In a real implementation we would check consumer lag or topic depth
            # kafka-python doesn't make it super easy to get "queue depth" without consuming
            # For this demo, we will return a placeholder or investigate admin client
            
            # Simulated data
            queues.append({
                "name": "agent-tasks",
                "tasks": [
                    {
                        "id": "task-mock-1",
                        "priority": {"level": "high", "waiting_since_mins": 5},
                        "invoked_by": "agent-a",
                        "waiting_since_mins": 5
                    }
                ]
            })
            
        except Exception as e:
            print(f"Error collecting Queue data: {e}")
            return []
            
        return queues
