oracle-monitor/collectors/requirements.txt
kubernetes==29.0.0
kafka-python==2.0.2
redis==5.0.1
requests==2.31.0
pydantic==2.6.1
jsonschema==4.21.1
supabase==2.3.0
python-dotenv==1.0.1
